Fix one missing #include and a couple of warnings.
git-svn-id: http://svn.boost.org/svn/boost/trunk@32736 b8fc166d-592f-0410-95f2-cb63ce0dd405

+++ boost/boost/archive/basic_binary_iprimitive.hpp
@@ -146,7 +146,7 @@ basic_binary_iprimitive<Archive, Elem, Tr>::load_binary(
         static_cast<Elem *>(address), 
         s
     );
-    if(count != s)
+    if(count != static_cast<std::size_t>(s))
         boost::throw_exception(
             archive_exception(archive_exception::stream_error)
         );
